requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.3
djangorestframework==3.12.2
pep8==1.7.1
pip-tools==5.4.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
